requirements.txt
fastapi
uvicorn
httpx
beautifulsoup4
pypdf
openai
python-multipart
python-dotenv
supabase
